+++ packages/components/accordion/__tests__/accordion.test.tsx
@@ -1,0 +1,10 @@
+import { describe, expect, it } from "vitest";
+
+describe("Accordion", () => {
+  it("test", () => {
+    expect(2).toEqual(2);
+  }); // it("should render correctly", () => {
+  //   const wrapper = render(<Accordion>hiii Accordion</Accordion>);
+  //   expect(() => wrapper.unmount()).not.toThrow();
+  // });
+});