+++ packages/components/autocomplete/__tests__/autocomplete.test.tsx
@@ -1,0 +1,11 @@
+import { describe, expect, it } from "vitest";
+
+describe("Autocomplete", () => {
+  it("test", () => {
+    expect(2).toEqual(2);
+  });
+  // it("should render correctly", () => {
+  //   const wrapper = render(<AutoComplete>hiii</AutoComplete>);
+  //   expect(() => wrapper.unmount()).not.toThrow();
+  // });
+});