+++ packages/components/button/__tests__/button.test.tsx
@@ -4,7 +4,7 @@ import { Button } from "../src";
 
 describe("Button", () => {
   it("should render correctly", () => {
-    const wrapper = render(<Button>hiii</Button>);
+    const wrapper = render(<Button>Button</Button>);
     expect(() => wrapper.unmount()).not.toThrow();
   });
 });