+++ packages/components/dialog/__tests__/dialog.test.tsx
@@ -1,0 +1,11 @@
+import { describe, expect, it } from "vitest";
+
+describe("Dialog", () => {
+  it("test", () => {
+    expect(2).toEqual(2);
+  });
+  // it("should render correctly", () => {
+  //   const wrapper = render(<Dialog>hiii</Dialog>);
+  //   expect(() => wrapper.unmount()).not.toThrow();
+  // });
+});