+++ packages/components/drawer/__tests__/drawer.test.tsx
@@ -1,6 +1,9 @@
-import { describe } from "vitest";
+import { describe, expect, it } from "vitest";
 
-describe("Divider", () => {
+describe("Drawer", () => {
+  it("test", () => {
+    expect(2).toEqual(2);
+  });
   // it("should render correctly", () => {
   //   const wrapper = render(<Drawer>hiii</Drawer>);
   //   expect(() => wrapper.unmount()).not.toThrow();