+++ packages/components/link/__tests__/link.test.tsx
@@ -1,0 +1,10 @@
+import { render } from "@testing-library/react";
+import { describe, expect, it } from "vitest";
+import { Link } from "../src";
+
+describe("Link", () => {
+  it("should render correctly", () => {
+    const wrapper = render(<Link href="/">hiii</Link>);
+    expect(() => wrapper.unmount()).not.toThrow();
+  });
+});