+++ packages/components/menu/__test__/menu.test.tsx
@@ -1,8 +1,0 @@
-import { describe } from "vitest";
-
-describe("Menu", () => {
-  // it("should render correctly", () => {
-  //   const wrapper = render(<Menu>hiii</Menu>);
-  //   expect(() => wrapper.unmount()).not.toThrow();
-  // });
-});