+++ packages/components/menu/__tests__/menu.test.tsx
@@ -1,0 +1,10 @@
+import { describe, expect, it } from "vitest";
+
+describe("Menu", () => {
+  it("test", () => {
+    expect(2).toEqual(2);
+  }); // it("should render correctly", () => {
+  //   const wrapper = render(<Menu>hiii</Menu>);
+  //   expect(() => wrapper.unmount()).not.toThrow();
+  // });
+});