+++ packages/components/popover/__test__/popover.test.tsx
@@ -1,8 +1,0 @@
-import { describe } from "vitest";
-
-describe("Popover", () => {
-  // it("should render correctly", () => {
-  //   const wrapper = render(<Popover>hiii</Popover>);
-  //   expect(() => wrapper.unmount()).not.toThrow();
-  // });
-});