+++ packages/components/popover/__tests__/popover.test.tsx
@@ -1,0 +1,11 @@
+import { describe, expect, it } from "vitest";
+
+describe("Popover", () => {
+  it("test", () => {
+    expect(2).toEqual(2);
+  });
+  // it("should render correctly", () => {
+  //   const wrapper = render(<Popover>hiii</Popover>);
+  //   expect(() => wrapper.unmount()).not.toThrow();
+  // });
+});