+++ packages/components/select/__tests__/select.test.tsx
@@ -1,0 +1,11 @@
+import { describe, expect, it } from "vitest";
+
+describe("Select", () => {
+  it("test", () => {
+    expect(2).toEqual(2);
+  });
+  // it("should render correctly", () => {
+  //   const wrapper = render(<Select>hiii</Select>);
+  //   expect(() => wrapper.unmount()).not.toThrow();
+  // });
+});