+++ packages/components/tab/__test__/tab.test.tsx
@@ -1,8 +1,0 @@
-import { describe } from "vitest";
-
-describe("Tab", () => {
-  // it("should render correctly", () => {
-  //   const wrapper = render(<Tab>hiii</Tab>);
-  //   expect(() => wrapper.unmount()).not.toThrow();
-  // });
-});