+++ packages/components/tab/__tests__/tab.test.tsx
@@ -1,0 +1,11 @@
+import { describe, expect, it } from "vitest";
+
+describe("Tab", () => {
+  it("test", () => {
+    expect(2).toEqual(2);
+  });
+  // it("should render correctly", () => {
+  //   const wrapper = render(<Tab>hiii</Tab>);
+  //   expect(() => wrapper.unmount()).not.toThrow();
+  // });
+});