+++ packages/utils/src/time.ts
@@ -1,3 +1,0 @@
-import { addMinutes } from "date-fns";
-
-export const fDateTime = (date: Date | string) => addMinutes(new Date(date), 2);