+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/react@0.0.10
+
 ## 0.1.9
 
 ### Patch Changes