+++ packages/components/dialog/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/dialog
 
+## 0.0.7
+
+### Patch Changes
+
+- base components
+- Updated dependencies
+  - @jamsr-ui/button@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes