+++ packages/react/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @jamsr-ui/react
 
+## 0.0.10
+
+### Patch Changes
+
+- base components
+- Updated dependencies
+  - @jamsr-ui/autocomplete@0.0.7
+  - @jamsr-ui/image-upload@0.0.7
+  - @jamsr-ui/typography@0.0.7
+  - @jamsr-ui/accordion@0.0.8
+  - @jamsr-ui/otp-input@0.0.7
+  - @jamsr-ui/checkbox@0.0.7
+  - @jamsr-ui/progress@0.0.7
+  - @jamsr-ui/skeleton@0.0.7
+  - @jamsr-ui/divider@0.0.7
+  - @jamsr-ui/popover@0.0.7
+  - @jamsr-ui/tooltip@0.0.7
+  - @jamsr-ui/avatar@0.0.7
+  - @jamsr-ui/button@0.0.7
+  - @jamsr-ui/dialog@0.0.7
+  - @jamsr-ui/header@0.0.7
+  - @jamsr-ui/ripple@0.0.7
+  - @jamsr-ui/select@0.0.7
+  - @jamsr-ui/switch@0.0.7
+  - @jamsr-ui/alert@0.0.7
+  - @jamsr-ui/badge@0.0.7
+  - @jamsr-ui/input@0.0.7
+  - @jamsr-ui/radio@0.0.7
+  - @jamsr-ui/card@0.0.7
+  - @jamsr-ui/chip@0.0.7
+  - @jamsr-ui/link@0.0.7
+  - @jamsr-ui/menu@0.0.7
+  - @jamsr-ui/tab@0.0.7
+
 ## 0.0.9
 
 ### Patch Changes